Add Signatures to a Combined Signature Model

This function takes a signature collection and a model as input and adds selected signatures to a combined signature model. The default output is a data.frame in the 'combined_signature_model' style data.frame format from the Sigverse package.

Usage

sig_combine(signatures, model, format = c("combined", "signature"))

Arguments

signatures

A signature collection, typically a list of data.frames where each data.frame represents a signature with specific attributes.

model

A named numeric vector representing the contribution of each signature to the combined model. The names correspond to the signatures, and the values represent their contributions. The sum of the values in this vector should be less than or equal to 1.

format

A character string indicating the output format. If "combined", the function returns a 'combined_signature_model' data.frame where each row represents a contribution for a particular channel from a single signature (duplicate channels are not collapsed). If "signature", the function returns the data in the sigverse signature format, representing a novel signature created by combining the signatures in the collection according to the ratios described by the model.

Value

A data.frame in the 'combined_signature_model' style containing the selected signatures and their modified fractions based on the model.

See also

Examples

library(sigstash)

# Load a signature collection
signatures <- sig_load("COSMIC_v3.3.1_SBS_GRCh38")

# Create a model that represents a mix of SBS1 (40%) and SBS2 (60%)
model <- c(SBS1 = 0.4, SBS2 = 0.6)

# Add selected signatures to the combined model
combined_signatures <- sig_combine(signatures, model)
print(combined_signatures)
#> # A tibble: 192 × 5
#>    signature type  channel  fraction fraction_original
#>    <chr>     <chr> <chr>       <dbl>             <dbl>
#>  1 SBS1      C>A   A[C>A]A 0.000350           0.000876
#>  2 SBS1      C>A   A[C>A]C 0.000888           0.00222 
#>  3 SBS1      C>A   A[C>A]G 0.0000719          0.000180
#>  4 SBS1      C>A   A[C>A]T 0.000506           0.00127 
#>  5 SBS1      C>G   A[C>G]A 0.000736           0.00184 
#>  6 SBS1      C>G   A[C>G]C 0.000475           0.00119 
#>  7 SBS1      C>G   A[C>G]G 0.0000467          0.000117
#>  8 SBS1      C>G   A[C>G]T 0.000451           0.00113 
#>  9 SBS1      C>T   A[C>T]A 0.00989            0.0247  
#> 10 SBS1      C>T   A[C>T]C 0.00246            0.00615 
#> # ℹ 182 more rows

# Visualise using sigvis